pyproject.toml
[project]
name = "ds-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.11.0",
    "openai>=1.95.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
